java-sdk-449 Implement Initialize/Shutdown on provider registration - moves handling of providers out of OpenFeatureAPI into dedicated ProviderRepository - cleans up temp map for named providers/member for default provider on successful initialization - prepares API shutdown method by creating one in the Provider Repository
