+++ src/main/java/dev/openfeature/sdk/OpenFeatureAPI.java
@@ -5,12 +5,9 @@ import dev.openfeature.sdk.internal.AutoCloseableReentrantReadWriteLock;
 import lombok.extern.slf4j.Slf4j;
 
 import javax.annotation.Nullable;
-import java.util.*;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.atomic.AtomicReference;
-import java.util.function.Consumer;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 
 /**
  * A global singleton which holds base configuration for the OpenFeature library.
@@ -22,13 +19,9 @@ public class OpenFeatureAPI {
     static AutoCloseableReentrantReadWriteLock hooksLock = new AutoCloseableReentrantReadWriteLock();
     static AutoCloseableReentrantReadWriteLock contextLock = new AutoCloseableReentrantReadWriteLock();
 
+    private final ProviderRepository providerRepository = new ProviderRepository();
     private final List<Hook> apiHooks;
-    private final Map<String, FeatureProvider> providers = new ConcurrentHashMap<>();
-    private final ExecutorService taskExecutor = Executors.newCachedThreadPool();
-    private final AtomicReference<FeatureProvider> initializingDefaultProvider = new AtomicReference<>();
-    private final Map<String, FeatureProvider> initializingNamedProviders = new ConcurrentHashMap<>();
 
-    private AtomicReference<FeatureProvider> defaultProvider = new AtomicReference<>(new NoOpProvider());
     private EvaluationContext evaluationContext;
 
     private OpenFeatureAPI() {
@@ -49,7 +42,7 @@ public class OpenFeatureAPI {
     }
 
     public Metadata getProviderMetadata() {
-        return defaultProvider.get().getMetadata();
+        return getProvider().getMetadata();
     }
 
     public Metadata getProviderMetadata(String clientName) {
@@ -90,10 +83,7 @@ public class OpenFeatureAPI {
      * Set the default provider.
      */
     public void setProvider(FeatureProvider provider) {
-        if (provider == null) {
-            throw new IllegalArgumentException("Provider cannot be null");
-        }
-        initializeProvider(provider);
+        providerRepository.setProvider(provider);
     }
 
     /**
@@ -103,94 +93,14 @@ public class OpenFeatureAPI {
      * @param provider   The provider to set.
      */
     public void setProvider(String clientName, FeatureProvider provider) {
-        if (provider == null) {
-            throw new IllegalArgumentException("Provider cannot be null");
-        }
-        initializeProvider(clientName, provider);
-    }
-
-    private void shutdownProvider(FeatureProvider provider) {
-        if (provider == null) {
-            return;
-        }
-        taskExecutor.submit(() -> {
-            try {
-                provider.shutdown();
-            } catch (Exception e) {
-                log.error("Exception when shutting down feature provider {}", provider.getClass().getName(), e);
-            }
-        });
-    }
-
-    private void initializeProvider(FeatureProvider provider) {
-        initializingDefaultProvider.set(provider);
-        initializeProvider(provider, this::updateDefaultProviderAfterInitialization);
-    }
-
-    private void initializeProvider(String clientName, FeatureProvider provider) {
-        initializingNamedProviders.put(clientName, provider);
-        initializeProvider(provider, newProvider -> updateProviderAfterInitialization(clientName, newProvider));
-    }
-
-    private void initializeProvider(FeatureProvider provider, Consumer<FeatureProvider> afterInitialization) {
-        taskExecutor.submit(() -> {
-            try {
-                provider.initialize();
-                afterInitialization.accept(provider);
-            } catch (Exception e) {
-                log.error("Exception when initializing feature provider {}", provider.getClass().getName(), e);
-            }
-        });
-    }
-
-    private void updateProviderAfterInitialization(String clientName, FeatureProvider newProvider) {
-        Optional
-                .ofNullable(initializingNamedProviders.get(clientName))
-                .filter(initializingProvider -> initializingProvider.equals(newProvider))
-                .ifPresent(provider -> updateNamedProviderAfterInitialization(clientName, provider));
-    }
-
-    private void updateDefaultProviderAfterInitialization(FeatureProvider initializedProvider) {
-        Optional
-                .ofNullable(this.initializingDefaultProvider.get())
-                .filter(initializingProvider -> initializingProvider.equals(initializedProvider))
-                .ifPresent(this::replaceDefaultProvider);
-    }
-
-    private void replaceDefaultProvider(FeatureProvider provider) {
-        FeatureProvider oldProvider = this.defaultProvider.getAndSet(provider);
-        if (isOldProviderNotBoundByName(oldProvider)) {
-            shutdownProvider(oldProvider);
-        }
-    }
-
-    private boolean isOldProviderNotBoundByName(FeatureProvider oldProvider) {
-        return !this.providers.containsValue(oldProvider);
-    }
-
-    private void updateNamedProviderAfterInitialization(String clientName, FeatureProvider initializedProvider) {
-        Optional
-                .ofNullable(this.initializingNamedProviders.get(clientName))
-                .filter(initializingProvider -> initializingProvider.equals(initializedProvider))
-                .ifPresent(provider -> replaceNamedProviderAndShutdownOldOne(clientName, provider));
-    }
-
-    private void replaceNamedProviderAndShutdownOldOne(String clientName, FeatureProvider provider) {
-        FeatureProvider oldProvider = this.providers.put(clientName, provider);
-        if (isOldProviderNotBoundMultipleTimes(oldProvider)) {
-            shutdownProvider(oldProvider);
-        }
-    }
-
-    private boolean isOldProviderNotBoundMultipleTimes(FeatureProvider oldProvider) {
-        return !(this.providers.containsValue(oldProvider) || this.defaultProvider.get().equals(oldProvider));
+        providerRepository.setProvider(clientName, provider);
     }
 
     /**
      * Return the default provider.
      */
     public FeatureProvider getProvider() {
-        return defaultProvider.get();
+        return providerRepository.getProvider();
     }
 
     /**
@@ -200,7 +110,7 @@ public class OpenFeatureAPI {
      * @return A named {@link FeatureProvider}
      */
     public FeatureProvider getProvider(String name) {
-        return Optional.ofNullable(name).map(this.providers::get).orElse(this.defaultProvider.get());
+        return providerRepository.getProvider(name);
     }
 
     /**
